Introduce HTTP OTLP exporter
The integration utilizes the HttpSender
interface introduced in OTel 1.28
and is based on the Vert.x HTTP Client
