+++ extensions/opentelemetry/runtime/src/main/java/io/quarkus/opentelemetry/runtime/exporter/otlp/OtlpRecorder.java
@@ -1,6 +1,7 @@
 package io.quarkus.opentelemetry.runtime.exporter.otlp;
 
 import static io.quarkus.opentelemetry.runtime.config.runtime.exporter.OtlpExporterRuntimeConfig.DEFAULT_GRPC_BASE_URI;
+import static io.quarkus.opentelemetry.runtime.config.runtime.exporter.OtlpExporterTracesConfig.Protocol.GRPC;
 import static io.quarkus.opentelemetry.runtime.config.runtime.exporter.OtlpExporterTracesConfig.Protocol.HTTP_PROTOBUF;
 
 import java.net.URI;
@@ -16,6 +17,8 @@ import jakarta.enterprise.util.TypeLiteral;
 
 import io.opentelemetry.api.metrics.MeterProvider;
 import io.opentelemetry.exporter.internal.ExporterBuilderUtil;
+import io.opentelemetry.exporter.internal.http.HttpExporter;
+import io.opentelemetry.exporter.internal.otlp.traces.TraceRequestMarshaler;
 import io.opentelemetry.exporter.otlp.internal.OtlpUserAgent;
 import io.opentelemetry.sdk.trace.export.BatchSpanProcessor;
 import io.opentelemetry.sdk.trace.export.BatchSpanProcessorBuilder;
@@ -41,12 +44,12 @@ public class OtlpRecorder {
             OTelRuntimeConfig otelRuntimeConfig,
             OtlpExporterRuntimeConfig exporterRuntimeConfig,
             TlsConfig tlsConfig, Supplier<Vertx> vertx) {
-        URI grpcBaseUri = getGrpcBaseUri(exporterRuntimeConfig); // do the creation and validation here in order to preserve backward compatibility
+        URI baseUri = getBaseUri(exporterRuntimeConfig); // do the creation and validation here in order to preserve backward compatibility
         return new Function<>() {
             @Override
             public LateBoundBatchSpanProcessor apply(
                     SyntheticCreationalContext<LateBoundBatchSpanProcessor> context) {
-                if (otelRuntimeConfig.sdkDisabled() || grpcBaseUri == null) {
+                if (otelRuntimeConfig.sdkDisabled() || baseUri == null) {
                     return RemoveableLateBoundBatchSpanProcessor.INSTANCE;
                 }
                 // Only create the OtlpGrpcSpanExporter if an endpoint was set in runtime config and was properly validated at startup
@@ -57,7 +60,7 @@ public class OtlpRecorder {
                 }
 
                 try {
-                    var spanExporter = createOtlpGrpcSpanExporter(exporterRuntimeConfig, vertx.get(), grpcBaseUri);
+                    var spanExporter = createSpanExporter(exporterRuntimeConfig, vertx.get(), baseUri);
 
                     BatchSpanProcessorBuilder processorBuilder = BatchSpanProcessor.builder(spanExporter);
 
@@ -73,102 +76,95 @@ public class OtlpRecorder {
                 }
             }
 
+            private SpanExporter createSpanExporter(OtlpExporterRuntimeConfig exporterRuntimeConfig,
+                    Vertx vertx, final URI baseUri) {
+                OtlpExporterTracesConfig tracesConfig = exporterRuntimeConfig.traces();
+                if (tracesConfig.protocol().isEmpty()) {
+                    throw new IllegalStateException("No OTLP protocol specified. " +
+                            "Please check `quarkus.otel.exporter.otlp.traces.protocol` property");
+                }
+
+                String protocol = tracesConfig.protocol().get();
+                if (GRPC.equals(protocol)) {
+                    return createOtlpGrpcSpanExporter(exporterRuntimeConfig, vertx, baseUri);
+                } else if (HTTP_PROTOBUF.equals(protocol)) {
+                    return createHttpSpanExporter(exporterRuntimeConfig, vertx, baseUri, protocol);
+                }
+
+                throw new IllegalArgumentException(String.format("Unsupported OTLP protocol %s specified. " +
+                        "Please check `quarkus.otel.exporter.otlp.traces.protocol` property", protocol));
+            }
+
             private SpanExporter createOtlpGrpcSpanExporter(OtlpExporterRuntimeConfig exporterRuntimeConfig,
-                    Vertx vertx, final URI grpcBaseUri) {
+                    Vertx vertx, final URI baseUri) {
 
                 OtlpExporterTracesConfig tracesConfig = exporterRuntimeConfig.traces();
-                if (tracesConfig.protocol().isPresent()) {
-                    if (!tracesConfig.protocol().get().equals(HTTP_PROTOBUF)) {
-                        throw new IllegalStateException("Only the GRPC Exporter is currently supported. " +
-                                "Please check `quarkus.otel.exporter.otlp.traces.protocol` property");
-                    }
-                }
-
-                boolean compressionEnabled = false;
-                if (tracesConfig.compression().isPresent()) {
-                    compressionEnabled = (tracesConfig.compression().get() == CompressionType.GZIP);
-                }
-
-                Map<String, String> headersMap = new HashMap<>();
-                OtlpUserAgent.addUserAgentHeader(headersMap::put);
-                if (tracesConfig.headers().isPresent()) {
-                    List<String> headers = tracesConfig.headers().get();
-                    if (!headers.isEmpty()) {
-                        for (String header : headers) {
-                            if (header.isEmpty()) {
-                                continue;
-                            }
-                            String[] parts = header.split("=", 2);
-                            String key = parts[0].trim();
-                            String value = parts[1].trim();
-                            headersMap.put(key, value);
-                        }
-                    }
-                }
 
                 return new VertxGrpcExporter(
                         "otlp", // use the same as OTel does
                         "span", // use the same as OTel does
                         MeterProvider::noop,
-                        grpcBaseUri,
-                        compressionEnabled,
+                        baseUri,
+                        determineCompression(tracesConfig),
                         tracesConfig.timeout(),
-                        headersMap,
-                        new Consumer<>() {
-                            @Override
-                            public void accept(HttpClientOptions options) {
-                                configureTLS(options);
-                            }
-
-                            private void configureTLS(HttpClientOptions options) {
-                                // TODO: this can reuse existing stuff when https://github.com/quarkusio/quarkus/pull/33228 is in
-                                options.setKeyCertOptions(toPemKeyCertOptions(tracesConfig));
-                                options.setPemTrustOptions(toPemTrustOptions(tracesConfig));
-
-                                if (VertxGrpcExporter.isHttps(grpcBaseUri)) {
-                                    options.setSsl(true);
-                                    options.setUseAlpn(true);
-                                }
-                                if (tlsConfig.trustAll) {
-                                    options.setTrustAll(true);
-                                    options.setVerifyHost(false);
-                                }
-                            }
-
-                            private KeyCertOptions toPemKeyCertOptions(OtlpExporterTracesConfig configuration) {
-                                PemKeyCertOptions pemKeyCertOptions = new PemKeyCertOptions();
-                                OtlpExporterTracesConfig.KeyCert keyCert = configuration.keyCert();
-                                if (keyCert.certs().isPresent()) {
-                                    for (String cert : keyCert.certs().get()) {
-                                        pemKeyCertOptions.addCertPath(cert);
-                                    }
-                                }
-                                if (keyCert.keys().isPresent()) {
-                                    for (String cert : keyCert.keys().get()) {
-                                        pemKeyCertOptions.addKeyPath(cert);
-                                    }
-                                }
-                                return pemKeyCertOptions;
-                            }
-
-                            private PemTrustOptions toPemTrustOptions(OtlpExporterTracesConfig configuration) {
-                                PemTrustOptions pemTrustOptions = new PemTrustOptions();
-                                OtlpExporterTracesConfig.TrustCert trustCert = configuration.trustCert();
-                                if (trustCert.certs().isPresent()) {
-                                    for (String cert : trustCert.certs().get()) {
-                                        pemTrustOptions.addCertPath(cert);
-                                    }
-                                }
-                                return pemTrustOptions;
-                            }
-                        },
+                        populateTracingExportHttpHeaders(tracesConfig),
+                        new HttpClientOptionsConsumer(tracesConfig, baseUri, tlsConfig),
                         vertx);
 
+            }
+
+            private SpanExporter createHttpSpanExporter(OtlpExporterRuntimeConfig exporterRuntimeConfig, Vertx vertx,
+                    URI baseUri, String protocol) {
+
+                OtlpExporterTracesConfig tracesConfig = exporterRuntimeConfig.traces();
+
+                boolean exportAsJson = false; //TODO: this will be enhanced in the future
+
+                return new VertxHttpExporter(new HttpExporter<TraceRequestMarshaler>(
+                        "otlp", // use the same as OTel does
+                        "span", // use the same as OTel does
+                        new VertxHttpExporter.VertxHttpSender(
+                                baseUri,
+                                determineCompression(tracesConfig),
+                                tracesConfig.timeout(),
+                                populateTracingExportHttpHeaders(tracesConfig),
+                                exportAsJson ? "application/json" : "application/x-protobuf",
+                                new HttpClientOptionsConsumer(tracesConfig, baseUri, tlsConfig),
+                                vertx),
+                        MeterProvider::noop,
+                        exportAsJson));
             }
         };
     }
 
-    private URI getGrpcBaseUri(OtlpExporterRuntimeConfig exporterRuntimeConfig) {
+    private static boolean determineCompression(OtlpExporterTracesConfig tracesConfig) {
+        if (tracesConfig.compression().isPresent()) {
+            return (tracesConfig.compression().get() == CompressionType.GZIP);
+        }
+        return false;
+    }
+
+    private static Map<String, String> populateTracingExportHttpHeaders(OtlpExporterTracesConfig tracesConfig) {
+        Map<String, String> headersMap = new HashMap<>();
+        OtlpUserAgent.addUserAgentHeader(headersMap::put);
+        if (tracesConfig.headers().isPresent()) {
+            List<String> headers = tracesConfig.headers().get();
+            if (!headers.isEmpty()) {
+                for (String header : headers) {
+                    if (header.isEmpty()) {
+                        continue;
+                    }
+                    String[] parts = header.split("=", 2);
+                    String key = parts[0].trim();
+                    String value = parts[1].trim();
+                    headersMap.put(key, value);
+                }
+            }
+        }
+        return headersMap;
+    }
+
+    private URI getBaseUri(OtlpExporterRuntimeConfig exporterRuntimeConfig) {
         String endpoint = resolveEndpoint(exporterRuntimeConfig).trim();
         if (endpoint.isEmpty()) {
             return null;
@@ -191,4 +187,62 @@ public class OtlpRecorder {
         return !DEFAULT_GRPC_BASE_URI.equals(endpoint);
     }
 
+    private static class HttpClientOptionsConsumer implements Consumer<HttpClientOptions> {
+        private final OtlpExporterTracesConfig tracesConfig;
+        private final URI baseUri;
+        private final TlsConfig tlsConfig;
+
+        public HttpClientOptionsConsumer(OtlpExporterTracesConfig tracesConfig, URI baseUri, TlsConfig tlsConfig) {
+            this.tracesConfig = tracesConfig;
+            this.baseUri = baseUri;
+            this.tlsConfig = tlsConfig;
+        }
+
+        @Override
+        public void accept(HttpClientOptions options) {
+            configureTLS(options);
+        }
+
+        private void configureTLS(HttpClientOptions options) {
+            // TODO: this can reuse existing stuff when https://github.com/quarkusio/quarkus/pull/33228 is in
+            options.setKeyCertOptions(toPemKeyCertOptions());
+            options.setPemTrustOptions(toPemTrustOptions());
+
+            if (OtlpExporterUtil.isHttps(baseUri)) {
+                options.setSsl(true);
+                options.setUseAlpn(true);
+            }
+            if (tlsConfig.trustAll) {
+                options.setTrustAll(true);
+                options.setVerifyHost(false);
+            }
+        }
+
+        private KeyCertOptions toPemKeyCertOptions() {
+            PemKeyCertOptions pemKeyCertOptions = new PemKeyCertOptions();
+            OtlpExporterTracesConfig.KeyCert keyCert = tracesConfig.keyCert();
+            if (keyCert.certs().isPresent()) {
+                for (String cert : keyCert.certs().get()) {
+                    pemKeyCertOptions.addCertPath(cert);
+                }
+            }
+            if (keyCert.keys().isPresent()) {
+                for (String cert : keyCert.keys().get()) {
+                    pemKeyCertOptions.addKeyPath(cert);
+                }
+            }
+            return pemKeyCertOptions;
+        }
+
+        private PemTrustOptions toPemTrustOptions() {
+            PemTrustOptions pemTrustOptions = new PemTrustOptions();
+            OtlpExporterTracesConfig.TrustCert trustCert = tracesConfig.trustCert();
+            if (trustCert.certs().isPresent()) {
+                for (String cert : trustCert.certs().get()) {
+                    pemTrustOptions.addCertPath(cert);
+                }
+            }
+            return pemTrustOptions;
+        }
+    }
 }
