+++ integration-tests/opentelemetry-vertx-exporter/src/main/java/io/quarkus/it/opentelemetry/vertx/exporter/HelloResource.java
@@ -1,0 +1,14 @@
+package io.quarkus.it.opentelemetry.vertx.exporter;
+
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+
+@Path("hello")
+public class HelloResource {
+
+    @GET
+    public String get() {
+        return "get";
+    }
+
+}
