+++ integration-tests/opentelemetry-vertx-exporter/src/main/resources/application.properties
