+++ integration-tests/opentelemetry-vertx-exporter/src/test/java/io/quarkus/it/opentelemetry/vertx/exporter/grpc/GrpcWithTLSNoCompressionTest.java
@@ -1,0 +1,13 @@
+package io.quarkus.it.opentelemetry.vertx.exporter.grpc;
+
+import io.quarkus.it.opentelemetry.vertx.exporter.AbstractExporterTest;
+import io.quarkus.it.opentelemetry.vertx.exporter.OtelCollectorLifecycleManager;
+import io.quarkus.test.common.QuarkusTestResource;
+import io.quarkus.test.common.ResourceArg;
+import io.quarkus.test.junit.QuarkusTest;
+
+@QuarkusTest
+@QuarkusTestResource(value = OtelCollectorLifecycleManager.class, initArgs = @ResourceArg(name = "enableTLS", value = "true"), restrictToAnnotatedClass = true)
+public class GrpcWithTLSNoCompressionTest extends AbstractExporterTest {
+
+}
