--- integration-tests/opentelemetry-vertx-grpc-exporter/src/main/java/io/quarkus/it/opentelemetry/vertx/grpc/exporter/HelloResource.java
@@ -1,14 +1,0 @@
-package io.quarkus.it.opentelemetry.vertx.grpc.exporter;
-
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-
-@Path("hello")
-public class HelloResource {
-
-    @GET
-    public String get() {
-        return "get";
-    }
-
-}
