--- integration-tests/opentelemetry-vertx-grpc-exporter/src/test/java/io/quarkus/it/opentelemetry/vertx/grpc/exporter/NoTLSNoCompressionTest.java
@@ -1,10 +1,0 @@
-package io.quarkus.it.opentelemetry.vertx.grpc.exporter;
-
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.junit.QuarkusTest;
-
-@QuarkusTest
-@QuarkusTestResource(value = OtelCollectorLifecycleManager.class, restrictToAnnotatedClass = true)
-public class NoTLSNoCompressionTest extends AbstractExporterTest {
-
-}
