--- integration-tests/opentelemetry-vertx-grpc-exporter/src/test/java/io/quarkus/it/opentelemetry/vertx/grpc/exporter/Traces.java
@@ -1,19 +1,0 @@
-package io.quarkus.it.opentelemetry.vertx.grpc.exporter;
-
-import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
-
-import io.opentelemetry.proto.collector.trace.v1.ExportTraceServiceRequest;
-
-public final class Traces {
-
-    private final List<ExportTraceServiceRequest> traceRequests = new CopyOnWriteArrayList<>();
-
-    public List<ExportTraceServiceRequest> getTraceRequests() {
-        return traceRequests;
-    }
-
-    public void reset() {
-        traceRequests.clear();
-    }
-}
