--- integration-tests/opentelemetry-vertx-grpc-exporter/src/test/resources/otel-config.yaml
@@ -1,43 +1,0 @@
-extensions:
-  health_check: {}
-receivers:
-  otlp:
-    protocols:
-      grpc:
-        endpoint: 0.0.0.0:4317
-#      http:
-#        endpoint: 0.0.0.0:4318
-  otlp/mtls:
-    protocols:
-      grpc:
-        endpoint: 0.0.0.0:5317
-        tls:
-          client_ca_file: $MTLS_CLIENT_CERTIFICATE
-          cert_file: $MTLS_SERVER_CERTIFICATE
-          key_file: $MTLS_SERVER_KEY
-#      http:
-#        endpoint: 0.0.0.0:5318
-#        tls:
-#          client_ca_file: $MTLS_CLIENT_CERTIFICATE
-#          cert_file: $MTLS_SERVER_CERTIFICATE
-#          key_file: $MTLS_SERVER_KEY
-exporters:
-  logging:
-    verbosity: $LOGGING_EXPORTER_VERBOSITY_LEVEL
-  otlp:
-    endpoint: $OTLP_EXPORTER_ENDPOINT
-    tls:
-      insecure: true
-    compression: none
-service:
-  extensions: [health_check]
-  pipelines:
-#    metrics:
-#      receivers: [otlp, otlp/mtls]
-#      exporters: [logging, otlp]
-    traces:
-      receivers: [otlp, otlp/mtls]
-      exporters: [logging, otlp]
-#    logs:
-#      receivers: [otlp, otlp/mtls]
-#      exporters: [logging, otlp]
