+++ integration-tests/pom.xml
@@ -358,7 +358,7 @@
                 <module>opentelemetry-vertx</module>
                 <module>opentelemetry-reactive</module>
                 <module>opentelemetry-grpc</module>
-                <module>opentelemetry-vertx-grpc-exporter</module>
+                <module>opentelemetry-vertx-exporter</module>
                 <module>opentelemetry-reactive-messaging</module>
                 <module>logging-json</module>
                 <module>jaxb</module>
